average, int to double

+++ libfractal/fractal.h
@@ -8,7 +8,7 @@ struct fractal {
     double a;
     double b;
     int* values;
-    int* average;
+    double* average;
 } fract;
 
 /*
